(testing): Initial Commit Regression Tests
Of course I'm still going to have tests baked into go(langs) testing
framework itself, but a good dotfile manager should have predictable
behaviour when used in an installation environment. That's what these
tests are.

I've setup an environment with rspec. It uses the built dotty
executable (from the directory above the tests) and a temporary
installation and configuration directory to run dotty on some configs
(in the config directory) and make sure the installation directory is
in the correct state.

Basically it tests each directive in an isolated environment. These
tests will also serve to make sure we know when we break backwards
compatability.