(bug): Fix #dot/link-gen can't access ctx env
For now I'm deferring variable evaluation in #dot/link-gen until the
:link directive itself is evaluated, this fixes the issue with
environment variables assigned in :def not being available when we parse
the tag and is thus replaced with the empty string and not attempted
again when we try a context that does have it assigned.

Unfortunately this isn't a perfect fix, for one it makes the code more
ugly (but it doesn't really need help with that atm). When I get a
chance I'm gonna reimplement dotty as a minimal lisp instead of just a
lispy configuration interpreter just to avoid this headache with types.